very rough draft of checkout and progress callback design general flow: global git_init and git_shutdown to start and stop the subsystem period.
for sessions:
gitObject for session data holders.
includes a progress struct for the requisite globally bound callbacks required by libgit2 with a hook-back to the object instance in the payload.
derives from ScriptTickObject so we can have it do periodic updates when checking out data with callbacks for start/stop/progress